pyproject.toml
[project]
name = "rfp-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "chromadb>=1.0.10",
    "files-to-prompt>=0.6",
    "fpdf2>=2.8.3",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.4.7",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dateutil>=2.9.0.post0",
    "reportlab>=4.4.1",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "tavily-python>=0.7.2",
    "tiktoken>=0.9.0",
    "uuid>=1.30",
]
